requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
python-dotenv
celery
redis
streamlit
streamlit-lottie
requests
pillow
openai
langchain-core